src/requirements.txt
avro==1.12.0
certifi==2025.4.26
charset-normalizer==3.4.1
greenlet==3.2.2
grpcio==1.71.0
grpcio-tools==1.71.0
idna==3.10
numpy==2.2.5
pandas==2.2.3
protobuf==5.29.4
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
